Recarrega resultados locais, se disponíveis

+++ spotipy/test-user-all-tracks.py
@@ -2,6 +2,7 @@
 # Dados obtidos são salvos em arquivos JSON, na pasta do script
 
 import json
+import os.path
 import spotipy
 import time
 from spotipy.oauth2 import SpotifyClientCredentials
@@ -20,76 +21,91 @@ sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager, auth
 sent_requests = 0
 
 # Obtém todos os artistas seguidos
-artists = []
-with open('artists.json', 'w') as fp:
-    results = sp.current_user_followed_artists(limit=50)
-    sent_requests += 1
-    artists.extend(results['artists']['items'])
-
-    while True:
-        print("Obtendo artistas...")
+if os.path.isfile('artists_final.json'):
+    # Recarrega resultados anteriores, se disponíveis
+    with open('artists_final.json', 'r') as fp:
+        artists = json.load(fp)
+else:
+    artists = []
+    with open('artists.json', 'w') as fp:
+        results = sp.current_user_followed_artists(limit=50)
+        sent_requests += 1
         artists.extend(results['artists']['items'])
-        json.dump(results['artists']['items'], fp)   # Salva resultado parcial
-        if results['artists']['next']:
-            if sent_requests > 200:
-                sent_requests = 0
-                print('Cooldown')
-                time.sleep(30)
 
-            results = sp.next(results['artists'])
-            sent_requests += 1
-        else:
-            break
+        while True:
+            print("Obtendo artistas...")
+            artists.extend(results['artists']['items'])
+            json.dump(results['artists']['items'], fp)   # Salva resultado parcial
+            if results['artists']['next']:
+                if sent_requests > 200:
+                    sent_requests = 0
+                    print('Cooldown')
+                    time.sleep(30)
+
+                results = sp.next(results['artists'])
+                sent_requests += 1
+            else:
+                break
             
-# Subtitui resultados parciais com resultados finais
-with open('artists.json', 'w') as fp:
-    json.dump(artists, fp)
+    # Salva resultado final
+    with open('artists_final.json', 'w') as fp:
+        json.dump(artists, fp)
 
 # Obtém todos os álbuns
-albums = []
-with open('albums.json', 'w') as fp:
-    for artist in artists:
-        print("Obtendo álbuns...")
-        results = sp.artist_albums(artist['id'], album_type='album', limit=50)
-        sent_requests += 1
-        while True:
-            albums.extend(results['items'])
-            json.dump(results['items'], fp)   # Salva resultado parcial
-            if results['next']:
-                if sent_requests > 200:
-                    sent_requests = 0
-                    print('Cooldown')
-                    time.sleep(30)
-                results = sp.next(results)
-            else:
-                break
+if os.path.isfile('albums_final.json'):
+    # Recarrega resultados anteriores, se disponíveis
+    with open('albums_final.json', 'r') as fp:
+        albums = json.load(fp)
+else:
+    albums = []
+    with open('albums.json', 'w') as fp:
+        for artist in artists:
+            print("Obtendo álbuns...")
+            results = sp.artist_albums(artist['id'], album_type='album', limit=50)
+            sent_requests += 1
+            while True:
+                albums.extend(results['items'])
+                json.dump(results['items'], fp)   # Salva resultado parcial
+                if results['next']:
+                    if sent_requests > 200:
+                        sent_requests = 0
+                        print('Cooldown')
+                        time.sleep(30)
+                    results = sp.next(results)
+                else:
+                    break
 
-# Subtitui resultados parciais com resultados finais
-with open('albums.json', 'w') as fp:
-    json.dump(albums, fp)
+    # Salva resultado final
+    with open('albums_final.json', 'w') as fp:
+        json.dump(albums, fp)
 
 # Obtém todas as faixas dos álbuns
-album_tracks = []
-with open('album_tracks.json', 'w') as fp:
-    for album in albums:
-        print("Obtendo faixas dos álbuns...")
-        results = sp.album_tracks(album['id'], limit=50)
-        sent_requests += 1
-        while True:
-            album_tracks.extend(results['items'])
-            json.dump(results['items'], fp)   # Salva resultado parcial
-            if results['next']:
-                if sent_requests > 200:
-                    sent_requests = 0
-                    print('Cooldown')
-                    time.sleep(30)
-                results = sp.next(results)
-            else:
-                break
+if os.path.isfile('album_tracks_final.json'):
+    # Recarrega resultados anteriores, se disponíveis
+    with open('album_tracks_final.json', 'r') as fp:
+        album_tracks = json.load(fp)
+else:
+    album_tracks = []
+    with open('album_tracks.json', 'w') as fp:
+        for album in albums:
+            print("Obtendo faixas dos álbuns...")
+            results = sp.album_tracks(album['id'], limit=50)
+            sent_requests += 1
+            while True:
+                album_tracks.extend(results['items'])
+                json.dump(results['items'], fp)   # Salva resultado parcial
+                if results['next']:
+                    if sent_requests > 200:
+                        sent_requests = 0
+                        print('Cooldown')
+                        time.sleep(30)
+                    results = sp.next(results)
+                else:
+                    break
 
-# Subtitui resultados parciais com resultados finais
-with open('album_tracks.json', 'w') as fp:
-    json.dump(album_tracks, fp)
+    # Salva resultado final
+    with open('album_tracks_final.json', 'w') as fp:
+        json.dump(album_tracks, fp)
 
 # Compila lista com ID das faixas
 track_ids = []
@@ -97,20 +113,25 @@ for track in album_tracks:
     track_ids.append(track['id'])
 
 # Obtém detalhes de todas as faixas
-tracks = []
-with open('tracks.json', 'w') as fp:
-    for i in range(0, len(track_ids), 50):
-        if sent_requests > 200:
-            sent_requests = 0
-            print('Cooldown')
-            time.sleep(30)
+if os.path.isfile('tracks_final.json'):
+    # Recarrega resultados anteriores, se disponíveis
+    with open('tracks_final.json', 'r') as fp:
+        tracks = json.load(fp)
+else:
+    tracks = []
+    with open('tracks.json', 'w') as fp:
+        for i in range(0, len(track_ids), 50):
+            if sent_requests > 200:
+                sent_requests = 0
+                print('Cooldown')
+                time.sleep(30)
 
-        print("Obtendo detalhes das faixas...")
-        results = sp.tracks(track_ids[i:i+50])
-        sent_requests += 1
-        tracks.extend(results['tracks'])
-        json.dump(results['tracks'], fp)   # Salva resultado parcial
+            print("Obtendo detalhes das faixas...")
+            results = sp.tracks(track_ids[i:i+50])
+            sent_requests += 1
+            tracks.extend(results['tracks'])
+            json.dump(results['tracks'], fp)   # Salva resultado parcial
 
-# Subtitui resultados parciais com resultados finais
-with open('tracks.json', 'w') as fp:
-    json.dump(tracks, fp)
+    # Salva resultado final
+    with open('tracks_final.json', 'w') as fp:
+        json.dump(tracks, fp)
